+++ sdk/identity/src/device_code_flow/device_code_responses.rs
