--- sdk/identity/src/responses/mod.rs
@@ -1,2 +1,0 @@
-mod refresh_token_response;
-pub use refresh_token_response::RefreshTokenResponse;
